Added captcha validation to the export database option.
The reasoning behind this is that rogue javascript can
technically automate the process of exporting the entries,
reading back the passwords from the disk and thus decrypt
all messages.

This should deter such attacks by requiring the user to
manually enter a captcha before being able to export the
unencrypted entries.

The added benefit of this prompt alerts the users that an
extraction attempt is being made and also informs them that
database exports are unencrypted.

Signed-off-by: leogx9r <leogatesx9r@protonmail.com>