Update FAQ with new information related to Pokemon's Subpoena
Due to the new breaking about Pokemon issuing a subpoena on Discord
to reveal user information, FAQ has been updated to include this
bit of information as well as a discussion on a privacy-conscious
subreddit.

Further elaborate the types of information leaks users may expose that
may have been used to reduce this.

Resolves issue #23 (https://gitlab.com/leogx9r/DiscordCrypt/issues/23)